retrieveAllPassengerAccountsFromDb adjusted
retrieveAllPassengerAccountsFromDb method in PassengerAccountController
has been adjusted to hexagonal architecture.

Logic responsible for mapping PassengerAccountEntities to Pages
has been placed In PassengerAccountStorageAdapter class.

+++ passenger-account-service/src/main/java/pl/flywithbookedseats/passengeraccountservice/api/passenger/dto/PagePassengerAccountDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record PagePassengerAccountDto(
         List<PassengerAccount> passengerAccounts,
-        Long currentPage,
-        Long totalPages,
+        Integer currentPage,
+        Integer totalPages,
         Long totalElements
 ) {}
